Converting components to tsx
